Agentpolis and geographtools related changes, graph importer and geograph tools versions changed to snapshot,
Former-commit-id: cff003bdf2cb5e469893d5f19a75af8d6b449a5b

+++ src/main/java/cz/cvut/fel/aic/amodsim/config/AmodsimConfig.java
@@ -8,7 +8,7 @@ import java.util.Map;
 import ninja.fido.config.GeneratedConfig;
 
 public class AmodsimConfig implements GeneratedConfig {
-  public Double vehicleSpeedInMeters;
+  public Integer vehicleSpeedInMeters;
 
   public Boolean useTripCache;
 
@@ -52,7 +52,7 @@ public class AmodsimConfig implements GeneratedConfig {
   }
 
   public AmodsimConfig fill(Map amodsimConfig) {
-    this.vehicleSpeedInMeters = (Double) amodsimConfig.get("vehicle_speed_in_meters");
+    this.vehicleSpeedInMeters = (Integer) amodsimConfig.get("vehicle_speed_in_meters");
     this.useTripCache = (Boolean) amodsimConfig.get("use_trip_cache");
     this.experimentName = (String) amodsimConfig.get("experiment_name");
     this.tripsFilename = (String) amodsimConfig.get("trips_filename");
